fix(permissions): fix issue due to uppercase vs lowecase strings (#79)
* fix permissions issue + refactor in favor of enum

* clean

* clean unneeded public static

* remove not used action param

* use checkPermission instead of hasAdminManagementPermissions

+++ cibseven-webclient-core/src/main/java/org/cibseven/webapp/rest/AdminService.java
@@ -11,8 +11,10 @@ import java.util.Map;
 import java.util.Optional;
 
 import org.cibseven.webapp.auth.CIBUser;
+import org.cibseven.webapp.auth.SevenResourceType;
 import org.cibseven.webapp.exception.SystemException;
 import org.cibseven.webapp.providers.BpmProvider;
+import org.cibseven.webapp.providers.PermissionConstants;
 import org.cibseven.webapp.providers.SevenProvider;
 import org.cibseven.webapp.rest.model.Authorization;
 import org.cibseven.webapp.rest.model.NewUser;
@@ -71,7 +73,7 @@ public class AdminService extends BaseService implements InitializingBean {
 	@Parameter(description = "Specifies the field to sort by") @RequestParam Optional<String> sortBy,
 	@Parameter(description = "Specifies the order of the sorting") @RequestParam Optional<String> sortOrder,
 			Locale loc, CIBUser user) {
-		hasAdminManagementPermissions(user, "read", "user", List.of("ALL", "READ"));
+		checkPermission(user, SevenResourceType.USER, PermissionConstants.READ_ALL);
 		return bpmProvider.findUsers(id, firstName, firstNameLike, lastName, lastNameLike, 
 				email, emailLike, memberOfGroup, memberOfTenant, idIn, firstResult, maxResults, sortBy, sortOrder, user);
 	}
@@ -92,7 +94,7 @@ public class AdminService extends BaseService implements InitializingBean {
 	public void createUser(
 			@RequestBody NewUser user,
 			Locale loc, CIBUser flowUser) {
-		hasAdminManagementPermissions(flowUser, "create", "user", List.of("ALL", "CREATE"));
+		checkPermission(flowUser, SevenResourceType.USER, PermissionConstants.CREATE_ALL);
 		bpmProvider.createUser(user, flowUser);
 	}
 
@@ -115,7 +117,7 @@ public class AdminService extends BaseService implements InitializingBean {
 			@Parameter(description = "User Id") @PathVariable String userId, 
 			@RequestBody User user, 
 			Locale loc, CIBUser flowUser) {
-		hasAdminManagementPermissions(flowUser, "update", "user", List.of("ALL", "UPDATE"));
+		checkPermission(flowUser, SevenResourceType.USER, PermissionConstants.UPDATE_ALL);
 		bpmProvider.updateUserProfile(userId, user, flowUser);
 	}
 	
@@ -138,7 +140,7 @@ public class AdminService extends BaseService implements InitializingBean {
 			@Parameter(description = "Group Id") @PathVariable String groupId, 
 			@Parameter(description = "User Id") @PathVariable String userId, 
 			Locale loc, CIBUser flowUser) {
-		hasAdminManagementPermissions(flowUser, "delete", "group", List.of("ALL", "DELETE"));
+		checkPermission(flowUser, SevenResourceType.GROUP, PermissionConstants.DELETE_ALL);
 		bpmProvider.addMemberToGroup(groupId, userId, flowUser);
 	}
 	
@@ -161,7 +163,7 @@ public class AdminService extends BaseService implements InitializingBean {
 			@Parameter(description = "Group Id") @PathVariable String groupId, 
 			@Parameter(description = "User Id") @PathVariable String userId, 
 			Locale loc, CIBUser flowUser) {
-		hasAdminManagementPermissions(flowUser, "delete", "group", List.of("ALL", "DELETE"));
+		checkPermission(flowUser, SevenResourceType.GROUP, PermissionConstants.DELETE_ALL);
 		bpmProvider.deleteMemberFromGroup(groupId, userId, flowUser);
 	}
 	
@@ -186,7 +188,7 @@ public class AdminService extends BaseService implements InitializingBean {
 			@Parameter(description = "User Id") @PathVariable String userId, 
 			@RequestBody Map<String, Object> data, 
 			Locale loc, CIBUser user) {
-		hasAdminManagementPermissions(user, "update", "user", List.of("ALL", "UPDATE"));
+		checkPermission(user, SevenResourceType.USER, PermissionConstants.UPDATE_ALL);
 		bpmProvider.updateUserCredentials(userId, data, user);
 	}
 
@@ -205,7 +207,7 @@ public class AdminService extends BaseService implements InitializingBean {
 	public void deleteUser(
 			@Parameter(description = "User Id") @PathVariable String userId, 
 			Locale loc, CIBUser user) {
-		hasAdminManagementPermissions(user, "delete", "user", List.of("ALL", "DELETE"));
+		checkPermission(user, SevenResourceType.USER, PermissionConstants.DELETE_ALL);
 		bpmProvider.deleteUser(userId, user);
 	}
 
@@ -231,7 +233,7 @@ public class AdminService extends BaseService implements InitializingBean {
 			@Parameter(description = "Specifies the maximum number of results to return"
 					+ "<br>Will return less results if there are no more results left") @RequestParam Optional<String> maxResults,		
 			Locale loc, CIBUser user) {
-		hasAdminManagementPermissions(user, "read", "group", List.of("ALL", "READ"));
+		checkPermission(user, SevenResourceType.GROUP, PermissionConstants.READ_ALL);
 		String decodedMember = member.map(value -> {
 			try {
 				return URLDecoder.decode(value, StandardCharsets.UTF_8.toString());
@@ -280,7 +282,7 @@ public class AdminService extends BaseService implements InitializingBean {
 	public void createGroup(
 			@RequestBody UserGroup group,
 			Locale loc, CIBUser user) {
-		hasAdminManagementPermissions(user, "create", "group", List.of("ALL", "CREATE"));
+		checkPermission(user, SevenResourceType.GROUP, PermissionConstants.CREATE_ALL);
 		bpmProvider.createGroup(group, user);
 	}
 
@@ -303,7 +305,7 @@ public class AdminService extends BaseService implements InitializingBean {
 			@Parameter(description = "Group Id") @PathVariable String groupId, 
 			@RequestBody UserGroup group, 
 			Locale loc, CIBUser user) {
-		hasAdminManagementPermissions(user, "update" ,"group", List.of("ALL", "UPDATE"));
+		checkPermission(user, SevenResourceType.GROUP, PermissionConstants.UPDATE_ALL);
 		bpmProvider.updateGroup(groupId, group, user);
 	}
 
@@ -322,7 +324,7 @@ public class AdminService extends BaseService implements InitializingBean {
 	public void deleteGroup(
 			@Parameter(description = "Group Id") @PathVariable String groupId, 
 			Locale loc, CIBUser user) {
-		hasAdminManagementPermissions(user, "delete", "group", List.of("ALL", "DELETE"));
+		checkPermission(user, SevenResourceType.GROUP, PermissionConstants.DELETE_ALL);
 		bpmProvider.deleteGroup(groupId, user);
 	}
 	
@@ -350,7 +352,7 @@ public class AdminService extends BaseService implements InitializingBean {
 			@Parameter(description = "Maximum number of results to return"
 					+ "<br>Will return less results if there are no more results left") @RequestParam Optional<String> maxResults,
 			Locale loc, CIBUser user) {
-		hasAdminManagementPermissions(user, "read", "authorization", List.of("ALL", "READ"));
+		checkPermission(user, SevenResourceType.AUTHORIZATION, PermissionConstants.READ_ALL);
 		return bpmProvider.findAuthorization(id, type, userIdIn, groupIdIn, resourceType, resourceId,
 				sortBy, sortOrder, firstResult, maxResults, user);
 	}
@@ -378,7 +380,7 @@ public class AdminService extends BaseService implements InitializingBean {
 	public ResponseEntity<Authorization> createAuthorization(
 			@RequestBody Authorization authorization,
 			Locale loc, CIBUser user) {
-		hasAdminManagementPermissions(user, "create", "authorization", List.of("ALL", "CREATE"));
+		checkPermission(user, SevenResourceType.AUTHORIZATION, PermissionConstants.CREATE_ALL);
 		return bpmProvider.createAuthorization(authorization, user);
 	}
 
@@ -407,7 +409,7 @@ public class AdminService extends BaseService implements InitializingBean {
 			@Parameter(description = "Authorization Id") @PathVariable String authorizationId, 
 			@RequestBody Map<String, Object> data, 
 			Locale loc, CIBUser user) {
-		hasAdminManagementPermissions(user, "update", "authorization", List.of("ALL", "UPDATE"));
+		checkPermission(user, SevenResourceType.AUTHORIZATION, PermissionConstants.UPDATE_ALL);
 		bpmProvider.updateAuthorization(authorizationId, data, user);
 	}
 
@@ -426,7 +428,7 @@ public class AdminService extends BaseService implements InitializingBean {
 	public void deleteAuthorization(
 			@Parameter(description = "Authorization Id") @PathVariable String authorizationId, 
 			Locale loc, CIBUser user) {
-		hasAdminManagementPermissions(user, "delete", "authorization", List.of("ALL", "DELETE"));
+		checkPermission(user, SevenResourceType.AUTHORIZATION, PermissionConstants.DELETE_ALL);
 		bpmProvider.deleteAuthorization(authorizationId, user);
 	}
